pyproject.toml
[project]
name = "self-improving-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.102.0",
    "google-generativeai>=0.8.4",
    "streamlit>=1.42.2",
]

[project.scripts]
self-improving-rag = "self_improving_rag:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
